Refactor: [BADA-378] 코드 책임 분리 및 디버깅 콘솔 삭제 #272
